pyproject.toml
[project]
name = "gated-tree-clip"
version = "0.1.0"
description = "Gated Tree Clip (仮)"
authors = [
    { name = "n4okins", email = "63445867+n4okins@users.noreply.github.com" }
]
dependencies = [
    "torch==2.4.0+cu121",
    "torchaudio==2.4.0+cu121",
    "torchvision==0.19.0+cu121",
    "tqdm>=4.64.1",
    "spacy>=3.7.5",
    "torchdata>=0.7.1",
    "pandas>=2.2.2",
    "dask[complete]>=2024.8.1",
    "matplotlib>=3.9.2",
    "vaex>=4.17.0",
    "ray>=2.34.0",
    "nvidia-dali-cuda120>=1.40.0",
    "nvidia-nvjpeg2k-cu12>=0.8.0.38",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "ruff>=0.5.5",
    "mypy>=1.11.0",
    "jupyterlab>=4",
    "python-dotenv>=1.0.1",
    "hydra-core>=1.3.2",
    "pydantic>=2.8.2",
    "torchinfo>=1.8.0",
    "transformers>=4.43.3",
    "open-clip-torch>=2.26.1",
    "ipywidgets>=8.1.3",
    "widgetsnbextension>=4.0.11",
    "jax[cuda12]>=0.4.30",
    "flax>=0.8.5",
    "utils @ git+https://github.com/n4okins/utils.git",
    "fugashi[unidic]>=1.3.2",
    "ipadic>=1.0.0",
    "img2dataset>=1.45.0",
    "pip>=24.2",
]

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[[tool.rye.sources]]
name = "google-jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
type = "find-links"

[tool.ruff]
line-length = 128

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
pythonpath = "src"

[tool.mypy]
exclude = [
    "ignores/"
]
ignore_missing_imports = true
disable_error_code = "import-untyped"

[tool.hatch.build.targets.wheel]
packages = ["src/gated_tree_clip"]
